+++ src/main/java/com/jagrosh/jdautilities/commandclient/examples/GuildlistCommand.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2017 John Grosh (jagrosh).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.jagrosh.jdautilities.commandclient.examples;
+
+import java.awt.Color;
+import java.util.concurrent.TimeUnit;
+import com.jagrosh.jdautilities.commandclient.Command;
+import com.jagrosh.jdautilities.commandclient.CommandEvent;
+import com.jagrosh.jdautilities.menu.pagination.Paginator;
+import com.jagrosh.jdautilities.menu.pagination.PaginatorBuilder;
+import com.jagrosh.jdautilities.waiter.EventWaiter;
+import net.dv8tion.jda.core.Permission;
+import net.dv8tion.jda.core.entities.ChannelType;
+import net.dv8tion.jda.core.exceptions.PermissionException;
+
+/**
+ *
+ * @author John Grosh (jagrosh)
+ */
+public class GuildlistCommand extends Command {
+
+    private final PaginatorBuilder pbuilder;
+    public GuildlistCommand(EventWaiter waiter)
+    {
+        this.name = "guildlist";
+        this.help = "shows the list of guilds the bot is on";
+        this.arguments = "[pagenum]";
+        this.botPermissions = new Permission[]{Permission.MESSAGE_EMBED_LINKS, Permission.MESSAGE_ADD_REACTION};
+        this.guildOnly = false;
+        this.ownerCommand = true;
+        pbuilder = new PaginatorBuilder().setColumns(1)
+                .setItemsPerPage(10)
+                .showPageNumbers(true)
+                .waitOnSinglePage(false)
+                .useNumberedItems(false)
+                .setFinalAction(m -> {
+                    try {
+                        m.clearReactions().queue();
+                    } catch(PermissionException ex) {
+                        m.delete().queue();
+                    }
+                })
+                .setEventWaiter(waiter)
+                .setTimeout(1, TimeUnit.MINUTES);
+    }
+
+    @Override
+    protected void execute(CommandEvent event) {
+        int page = 1;
+        if(!event.getArgs().isEmpty())
+        {
+            try
+            {
+                page = Integer.parseInt(event.getArgs());
+            }
+            catch(NumberFormatException e)
+            {
+                event.reply(event.getClient().getError()+" `"+event.getArgs()+"` is not a valid integer!");
+                return;
+            }
+        }
+        pbuilder.setItems(new String[0]);
+        event.getJDA().getGuilds().stream()
+                .map(g -> "**"+g.getName()+"** (ID:"+g.getId()+") ~ "+g.getMembers().size()+" Members")
+                .forEach(s -> pbuilder.addItems(s));
+        Paginator p = pbuilder.setColor(event.isFromType(ChannelType.TEXT) ? event.getSelfMember().getColor() : Color.black)
+                .setText(event.getClient().getSuccess()+" Guilds that **"+event.getSelfUser().getName()+"** is connected to"
+                        +(event.getJDA().getShardInfo()==null ? ":" : "(Shard ID "+event.getJDA().getShardInfo().getShardId()+"):"))
+                .setUsers(event.getAuthor())
+                .build();
+        p.paginate(event.getChannel(), page);
+    }
+    
+}
