--- src/main/java/me/jagrosh/jdautilities/JDAUtilitiesInfo.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package me.jagrosh.jdautilities;
-
-/**
- *
- * @author John Grosh (john.a.grosh@gmail.com)
- */
-public class JDAUtilitiesInfo {
-    public static final String VERSION_MAJOR = "1";
-    public static final String VERSION_MINOR = "0";
-    public static final String VERSION = VERSION_MAJOR+"."+VERSION_MINOR;
-    public static final String GITHUB = "https://github.com/jagrosh/JDA-Utilities";
-    public static final String AUTHOR = "jagrosh";
-}
