--- src/main/java/me/jagrosh/jdautilities/commandclient/CommandListener.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2016 John Grosh (jagrosh).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package me.jagrosh.jdautilities.commandclient;
-
-import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
-
-/**
- *
- * @author John Grosh (jagrosh)
- */
-public interface CommandListener {
-    
-    public void onCommand(CommandEvent event, Command command);
-    
-    public void onCompletedCommand(CommandEvent event, Command command);
-    
-    public void onTerminatedCommand(CommandEvent event, Command command);
-    
-    public void onNonCommandMessage(MessageReceivedEvent event);
-}
