--- src/main/java/me/jagrosh/jdautilities/commandclient/examples/PingCommand.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2016 John Grosh (jagrosh).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package me.jagrosh.jdautilities.commandclient.examples;
-
-import java.time.temporal.ChronoUnit;
-import me.jagrosh.jdautilities.commandclient.Command;
-import me.jagrosh.jdautilities.commandclient.CommandEvent;
-
-/**
- *
- * @author John Grosh (jagrosh)
- */
-public class PingCommand extends Command {
-
-    public PingCommand()
-    {
-        this.name = "ping";
-        this.help = "checks the bot's latency";
-        this.guildOnly = false;
-        this.aliases = new String[]{"pong"};
-    }
-
-    @Override
-    protected void execute(CommandEvent event) {
-        event.getChannel().sendMessage("Ping: ...").queue(m -> {
-            m.editMessage("Ping: "+event.getMessage().getCreationTime().until(m.getCreationTime(), ChronoUnit.MILLIS)+"ms").queue();
-        });
-    }
-    
-}
